+++ src/server/utils/W3Helper.js
@@ -2,7 +2,7 @@ import fetch from 'cross-fetch'
 import md5 from 'md5'
 import conf from '../server.config'
 import logger from '../../imports/logger'
-import requestTimeout from '../utils/timeout'
+import requestTimeout from '../utils/async'
 
 const log = logger.child({ from: 'W3Helper' })
